Extract mock adapter tests to shareable module
Previously a set of behavioural tests for MockAdapter were introduced.
However, these tests cover behaviour that should be shared by all adapters,
and provides coverage that the other adapters don't yet have.

This change extracts the tests to a module that can be included in the
existing and any new adapter tests, as well as used by anyone developing their
own adapter.

Clients only need to define an `#adapter` method in their test class;
including the module then imports all the tests.
